+++ src/settings/Models/FinePolicy/FinePolicy.js
@@ -4,17 +4,19 @@ import {
 
 import { OverdueFine, Metadata } from '../common';
 import { FINE_POLICY_PATH } from '../../../constants/Validation/fine-policy';
+import ReminderFeesPolicy from '../ReminderFeesPolicy';
 
 export default class FinePolicy {
   static defaultFinePolicy() {
     return {
-      'countClosed': true,
-      'forgiveOverdueFine': true,
-      'gracePeriodRecall': true,
-      'overdueFine': { quantity: 0, intervalId: '' },
-      'maxOverdueFine': 0,
-      'maxOverdueRecallFine': 0,
-      'overdueRecallFine': { quantity: 0, intervalId: '' },
+      countClosed: true,
+      forgiveOverdueFine: true,
+      gracePeriodRecall: true,
+      overdueFine: { quantity: 0, intervalId: '' },
+      maxOverdueFine: 0,
+      maxOverdueRecallFine: 0,
+      overdueRecallFine: { quantity: 0, intervalId: '' },
+      reminderFeesPolicy: ReminderFeesPolicy.defaultReminderFeesPolicy(),
     };
   }
 
@@ -30,6 +32,7 @@ export default class FinePolicy {
     this.gracePeriodRecall = policy.gracePeriodRecall;
     this.maxOverdueRecallFine = policy.maxOverdueRecallFine;
     this.metadata = new Metadata(policy.metadata);
+    this.reminderFeesPolicy = new ReminderFeesPolicy(policy.reminderFeesPolicy);
   }
 
   hasValue(pathToValue) {