+++ test/jest/fixtures/reminderFeesPolicy.js
@@ -1,0 +1,13 @@
+const reminderFeesPolicy = {
+  reminderSchedule: [
+    {
+      interval: 2,
+      timeUnitId: 'day',
+      reminderFee: 1,
+      noticeMethodId: 'email',
+      noticeTemplateId: 'c4f23e08-3b0d-4424-afc3-1e228240a8b0'
+    },
+  ],
+};
+
+export default reminderFeesPolicy;